# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenMuara
  description: Local-first billing and payment virtualization layer.
  version: 0.1.1
  contact:
    name: OpenMuara
servers:
  - url: http://127.0.0.1:9000
    description: Local development server
paths:
...
  /_admin/transactions:
    get:
      summary: List recent transactions
      operationId: listTransactions
      responses:
        '200':
          description: Paginated list of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTransactions'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /fawry/charge:
    post:
      summary: Fawry-style charge request
      operationId: fawryCharge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FawryChargeRequest'
      responses:
        '200':
          description: Charge recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FawryChargeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    PaginatedTransactions:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            trace_id:
              type: string
    TransactionItem:
      type: object
      properties:
        itemCode:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
    Transaction:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        type:
          type: string
        amount:
          type: number
          format: double
        currency:
          type: string
        status:
          type: string
          enum: [new, paid, unpaid, refunded]
        customerRef:
          type: string
        idempotencyKey:
          type: string
        reference:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItem'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FawryChargeItem:
      type: object
      properties:
        itemId:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
    FawryChargeRequest:
      type: object
      required: [merchantCode, merchantRefNum, chargeItems, returnUrl, signature]
      properties:
        merchantCode:
          type: string
        merchantRefNum:
          type: string
        customerEmail:
          type: string
        customerName:
          type: string
        customerProfileId:
          type: string
        paymentExpiry:
          type: integer
          format: int64
        language:
          type: string
        chargeItems:
          type: array
          items:
            $ref: '#/components/schemas/FawryChargeItem'
        returnUrl:
          type: string
        signature:
          type: string
    FawryChargeResponse:
      type: object
      properties:
        status:
          type: string
        reference:
          type: string
...